Add files via upload
This project revolves around analyzing sales data for an e-commerce company to derive valuable insights into customer behavior, product performance, and overall business performance. Here's a breakdown of some key components:

Objectives:
The main objective of this project is to gain meaningful insights from the sales data. This includes analyzing customer behavior, evaluating product performance, and assessing overall business success.

Tasks and Techniques:
This project incorporates various tasks and techniques to achieve the objectives, which are:
•	Data Import: The dataset is imported into an SQL database for efficient management and analysis.
•	Exploratory Data Analysis: SQL queries are employed to retrieve crucial information such as the total number of orders, average order, best and worst performing product and overall revenue. Excel is then used to create visualizations like charts and graphs to represent the hourly and daily revenue trend, revenue by product category, and the top and bottom 5 brands based on total quantity ordered.

Conclusion and Recommendations: This project concludes with summarizing the key findings derived from the analysis and providing recommendations to enhance business performance based on the insights gained.

Skills and Tools:
Throughout the project, the two primary tools utilized are SQL and Excel. SQL is employed for data retrieval, manipulation, and analysis from the database, while Excel is used for data visualization, creating charts and graphs, and conducting further analysis.